bpf: use hwaddr for identification

+++ db/models.go
@@ -9,9 +9,9 @@ import (
 )
 
 type Usage struct {
-	Hardwareaddr int32
+	Hardwareaddr int64
 	Starttime    pgtype.Timestamp
 	Stoptime     pgtype.Timestamp
-	Egress       int32
-	Ingress      int32
+	Egress       int64
+	Ingress      int64
 }
